+++ radio/src/gui/128x64/menus.h
@@ -45,7 +45,7 @@ inline bool isModelMenuDisplayed()
 void onMainViewMenu(const char * result);
 void menuFirstCalib(event_t event);
 void menuMainView(event_t event);
-void menuViewTelemetryFrsky(event_t event);
+void menuViewTelemetry(event_t event);
 void menuSpecialFunctions(event_t event, CustomFunctionData * functions, CustomFunctionsContext * functionsContext);
 
 enum MenuRadioIndexes
