+++ radio/src/gui/128x64/view_telemetry.cpp
@@ -206,79 +206,3 @@ bool displayTelemetryScreen()
 
   return true;
 }
-
-enum NavigationDirection {
-  none,
-  up,
-  down
-};
-#define decrTelemetryScreen() direction = up
-#define incrTelemetryScreen() direction = down
-
-#if defined(NAVIGATION_XLITE)
-  #define EVT_KEY_PREVIOUS_VIEW(evt)         (evt == EVT_KEY_LONG(KEY_LEFT) && IS_SHIFT_PRESSED())
-  #define EVT_KEY_NEXT_VIEW(evt)             (evt == EVT_KEY_LONG(KEY_RIGHT) && IS_SHIFT_PRESSED())
-#elif defined(NAVIGATION_X7)
-  #define EVT_KEY_PREVIOUS_VIEW(evt)         (evt == EVT_KEY_LONG(KEY_PAGE))
-  #define EVT_KEY_NEXT_VIEW(evt)             (evt == EVT_KEY_BREAK(KEY_PAGE))
-#elif defined(NAVIGATION_9X)
-  #define EVT_KEY_PREVIOUS_VIEW(evt)         (evt == EVT_KEY_LONG(KEY_UP))
-  #define EVT_KEY_NEXT_VIEW(evt)             (evt == EVT_KEY_LONG(KEY_DOWN))
-#else
-  #define EVT_KEY_PREVIOUS_VIEW(evt)         (evt == EVT_KEY_FIRST(KEY_UP))
-  #define EVT_KEY_NEXT_VIEW(evt)             (evt == EVT_KEY_FIRST(KEY_DOWN))
-#endif
-
-void menuViewTelemetryFrsky(event_t event)
-{
-  enum NavigationDirection direction = none;
-
-  if (event == EVT_KEY_FIRST(KEY_EXIT) && TELEMETRY_SCREEN_TYPE(s_frsky_view) != TELEMETRY_SCREEN_TYPE_SCRIPT) {
-    killEvents(event);
-    chainMenu(menuMainView);
-  }
-#if defined(LUA)
-  else if (event == EVT_KEY_LONG(KEY_EXIT)) {
-    killEvents(event);
-    chainMenu(menuMainView);
-  }
-#endif
-  else if (EVT_KEY_PREVIOUS_VIEW(event)) {
-    killEvents(event);
-    decrTelemetryScreen();
-  }
-  else if (EVT_KEY_NEXT_VIEW(event)) {
-    killEvents(event);
-    incrTelemetryScreen();
-  }
-  else if (event == EVT_KEY_LONG(KEY_ENTER)) {
-    killEvents(event);
-    POPUP_MENU_ADD_ITEM(STR_RESET_TELEMETRY);
-    POPUP_MENU_ADD_ITEM(STR_RESET_FLIGHT);
-    POPUP_MENU_START(onMainViewMenu);
-  }
-
-  for (int i=0; i<=TELEMETRY_SCREEN_TYPE_MAX; i++) {
-    if (direction == up) {
-      if (s_frsky_view-- == 0)
-        s_frsky_view = TELEMETRY_VIEW_MAX;
-    }
-    else if (direction == down) {
-      if (s_frsky_view++ == TELEMETRY_VIEW_MAX)
-        s_frsky_view = 0;
-    }
-    else {
-      direction = down;
-    }
-    if (displayTelemetryScreen()) {
-      return;
-    }
-  }
-
-  drawTelemetryTopBar();
-  lcdDrawText(2*FW, 3*FH, "No Telemetry Screens");
-  displayRssiLine();
-}
-
-#undef EVT_KEY_PREVIOUS_VIEW
-#undef EVT_KEY_NEXT_VIEW
