+++ radio/src/gui/212x64/menus.h
@@ -39,7 +39,7 @@ void menuFirstCalib(event_t event);
 void menuMainViewChannelsMonitor(event_t event);
 void menuChannelsView(event_t event);
 void menuMainView(event_t event);
-void menuViewTelemetryFrsky(event_t event);
+void menuViewTelemetry(event_t event);
 void menuSpecialFunctions(event_t event, CustomFunctionData * functions, CustomFunctionsContext * functionsContext);
 void menuModelNotes(event_t event);
 void menuStatisticsView(event_t event);
