+++ radio/src/gui/212x64/view_telemetry.cpp
@@ -200,62 +200,3 @@ bool displayTelemetryScreen()
   return true;
 }
 
-enum NavigationDirection {
-  none,
-  up,
-  down
-};
-
-#define decrTelemetryScreen() direction = up
-#define incrTelemetryScreen() direction = down
-
-void menuViewTelemetryFrsky(event_t event)
-{
-  enum NavigationDirection direction = none;
-
-  if (event == EVT_KEY_FIRST(KEY_EXIT) && TELEMETRY_SCREEN_TYPE(s_frsky_view) != TELEMETRY_SCREEN_TYPE_SCRIPT) {
-    killEvents(event);
-    chainMenu(menuMainView);
-  }
-#if defined(LUA)
-  else if (event == EVT_KEY_LONG(KEY_EXIT)) {
-    killEvents(event);
-    chainMenu(menuMainView);
-  }
-#endif
-  else if (event == EVT_KEY_LONG(KEY_PAGE)) {
-    killEvents(event);
-    decrTelemetryScreen();
-  }
-  else if (event == EVT_KEY_BREAK(KEY_PAGE)) {
-    killEvents(event);
-    incrTelemetryScreen();
-  }
-  else if (event == EVT_KEY_LONG(KEY_ENTER)) {
-    killEvents(event);
-    POPUP_MENU_ADD_ITEM(STR_RESET_TELEMETRY);
-    POPUP_MENU_ADD_ITEM(STR_RESET_FLIGHT);
-    POPUP_MENU_START(onMainViewMenu);
-  }
-
-  for (int i=0; i<=TELEMETRY_SCREEN_TYPE_MAX; i++) {
-    if (direction == up) {
-      if (s_frsky_view-- == 0)
-        s_frsky_view = TELEMETRY_VIEW_MAX;
-    }
-    else if (direction == down) {
-      if (s_frsky_view++ == TELEMETRY_VIEW_MAX)
-        s_frsky_view = 0;
-    }
-    else {
-      direction = down;
-    }
-    if (displayTelemetryScreen()) {
-      return;
-    }
-  }
-
-  drawTelemetryTopBar();
-  lcdDrawText(8*FW, 3*FH, "No Telemetry Screens");
-  displayRssiLine();
-}
