+++ radio/src/gui/common/stdlcd/CMakeLists.txt
@@ -15,6 +15,7 @@ set(GUI_SRC
   ../common/stdlcd/radio_hardware.cpp
   ../common/stdlcd/radio_version.cpp
   ../common/stdlcd/radio_trainer.cpp
+  ../common/stdlcd/view_telemetry.cpp
   model_curve_edit.cpp
   )
